requirements.txt
psycopg2-binary==2.9.5 
pandas==1.5.2
tqdm==4.66.1